@@ -4,20 +4,18 @@
   "version": "1.1.0",
   "license": "MIT",
   "scripts": {
-    "dev": "run-p watch:tailwind sapper:dev",
-    "sapper:dev": "sapper dev --port 4400",
-    "export": "sapper export --legacy",
-    "start": "PORT=4400 node __sapper__/build",
-    "cy:run": "cypress run",
-    "cy:open": "cypress open",
-    "test": "run-p --race dev cy:run",
-    "watch:tailwind": "postcss static/tailwind.css -o static/index.css -w",
-    "build:tailwind": "cross-env NODE_ENV=production postcss static/tailwind.css -o static/index.css",
-    "build": "run-p build:tailwind buildS",
-    "buildS": "cross-env NODE_ENV=production sapper build --legacy",
+    "dev": "sapper dev --port 4400",
+    "build": "cross-env NODE_ENV=production sapper build --legacy",
+    "export": "cross-env NODE_ENV=production sapper export --legacy",
+    "start": "cross-env NODE_ENV=production node __sapper__/build",
     "live": "run-s build copy liveS",
     "liveS": "node devops/live.js",
-    "copy": "node devops/copy.js"
+    "format:check": "prettier --check './**/*.{js,ts,css,html,svelte}'",
+    "copy": "node devops/copy.js",
+    "cy:run": "cypress run",
+    "cy:open": "cypress open",
+    "dist": "cross-env NODE_ENV=production http-server __sapper__/export -p 4400 -P https://api.litekart.in",
+    "test": "run-p --race dev cy:run"
   },
   "homepage": "https://github.com/itswadesh/sapper-ecommerce#readme",
   "repository": {
@@ -28,37 +26,50 @@
     "url": "https://github.com/itswadesh/sapper-ecommerce/issues"
   },
   "dependencies": {
-    "@beyonk/svelte-carousel": "^2.1.1",
-    "@fullhuman/postcss-purgecss": "^1.3.0",
     "compression": "^1.7.1",
-    "cookie-universal": "^2.1.0",
     "http-proxy-middleware": "^0.20.0",
     "node-fetch": "^2.6.0",
     "polka": "0.5.2",
-    "siema": "^1.5.1",
     "sirv": "^0.4.0"
   },
   "devDependencies": {
     "@babel/core": "^7.8.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.0",
-    "@babel/plugin-transform-runtime": "^7.8.0",
-    "@babel/preset-env": "^7.8.2",
-    "@babel/runtime": "^7.8.0",
-    "cross-env": "^6.0.3",
+    "@babel/plugin-syntax-dynamic-import": "7.8.3",
+    "@babel/plugin-transform-runtime": "7.9.0",
+    "@babel/preset-env": "7.9.5",
+    "@babel/runtime": "7.9.2",
+    "@beyonk/svelte-carousel": "^2.4.0",
+    "@fullhuman/postcss-purgecss": "^2.1.2",
+    "@rollup/plugin-commonjs": "11.0.2",
+    "@rollup/plugin-node-resolve": "7.1.3",
+    "@rollup/plugin-replace": "2.3.2",
+    "autoprefixer": "9.7.6",
+    "cookie-universal": "^2.1.3",
+    "cross-env": "7.0.2",
+    "cssnano": "4.1.10",
     "dotenv": "^8.2.0",
-    "npm-run-all": "^4.1.5",
-    "postcss-cli": "^7.1.0",
-    "rollup": "^1.29.0",
-    "rollup-plugin-babel": "^4.0.2",
-    "rollup-plugin-commonjs": "^10.0.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-replace": "^2.0.0",
-    "rollup-plugin-svelte": "^5.1.1",
-    "rollup-plugin-terser": "^5.2.0",
-    "sapper": "^0.27.0",
+    "eslint": "^6.7.2",
+    "eslint-config-prettier": "^6.7.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-promise": "^4.2.1",
+    "npm-run-all": "4.1.5",
+    "postcss": "7.0.27",
+    "postcss-import": "12.0.1",
+    "postcss-load-config": "2.1.0",
+    "postcss-url": "8.0.0",
+    "prettier": "^2.0.5",
+    "rollup": "2.7.2",
+    "rollup-plugin-babel": "4.4.0",
+    "rollup-plugin-svelte": "5.2.1",
+    "rollup-plugin-terser": "5.3.0",
+    "sapper": "0.27.12",
     "shelljs": "^0.8.3",
     "ssh2shell": "^1.9.3",
-    "svelte": "^3.16.7",
-    "tailwindcss": "^1.1.4"
+    "svelte": "3.21.0",
+    "svelte-feather-icons": "^3.2.2",
+    "svelte-preprocess": "3.7.4",
+    "tailwindcss": "1.3.5"
   }
 }
