@@ -1,19 +1,19 @@
-import { sineOut } from "svelte/easing";
+import { sineOut } from 'svelte/easing'
 
-let duration = 250;
-let delay = duration;
+let duration = 250
+let delay = duration
 
-let delayZero = 0;
+let delayZero = 0
 
-export const fadeIn = _ => ({
-    duration,
-    delay,
-    easing: sineOut,
-    css: t => `opacity: ${t}`
-});
-export const fadeOut = _ => ({
-    duration,
-    delayZero,
-    easing: sineOut,
-    css: t => `opacity: ${t}`
-});
+export const fadeIn = (_) => ({
+  duration,
+  delay,
+  easing: sineOut,
+  css: (t) => `opacity: ${t}`,
+})
+export const fadeOut = (_) => ({
+  duration,
+  delayZero,
+  easing: sineOut,
+  css: (t) => `opacity: ${t}`,
+})
