@@ -1,42 +1,42 @@
-import { get, post } from "./../lib/api";
+import { get, post } from './../lib/api'
 async function authenticationMiddleware(req, res, next) {
   let user = null,
     cart = null,
-    settings = null;
-  const cookies = require("cookie-universal")(req, res);
-  if (cookies.get("token")) {
+    settings = null
+  const cookies = require('cookie-universal')(req, res)
+  if (cookies.get('token')) {
     try {
-      user = await get("users/me", null, cookies.get("token"));
+      user = await get('users/me', null, cookies.get('token'))
     } catch (e) {
-      console.log("err at users", e.toString());
+      console.log('err at users', e.toString())
     }
-    req.user = user;
-    req.token = cookies.get("token");
+    req.user = user
+    req.token = cookies.get('token')
   } else {
-    req.user = {};
-    req.token = null;
-    cookies.set("token", null);
+    req.user = {}
+    req.token = null
+    cookies.set('token', null)
   }
   try {
     cart = await get(
-      "cart",
+      'cart',
       null,
-      cookies.get("token"),
-      `guest=${cookies.get("guest")}`
-    );
+      cookies.get('token'),
+      `guest=${cookies.get('guest')}`
+    )
   } catch (e) {
-    req.user = {};
-    req.token = null;
-    cookies.set("token", null);
-    console.log("err at cart", e.toString());
+    req.user = {}
+    req.token = null
+    cookies.set('token', null)
+    console.log('err at cart', e.toString())
   }
-  req.cart = cart;
+  req.cart = cart
   try {
-    settings = (await get("settings", null, cookies.get("token"))).data;
+    settings = (await get('settings', null, cookies.get('token'))).data
   } catch (e) {
-    console.log("err at settings", e.toString());
+    console.log('err at settings', e.toString())
   }
-  req.settings = settings;
-  next();
+  req.settings = settings
+  next()
 }
-export { authenticationMiddleware };
+export { authenticationMiddleware }
