@@ -1,13 +1,13 @@
 <script>
-  import Header from "./../components/Header.svelte";
-  import Hero from "./../components/Hero.svelte";
-  import Banners from "./../components/Banners.svelte";
-  import Deals from "./../components/Deals.svelte";
-  import Carousel from "./../components/Carousel.svelte";
-  import ProductSkeleton from "./../components/ProductSkeleton.svelte";
-  import { get, put, post } from "./../lib/api";
-  import { HOST } from "./../config";
-  import { goto } from "@sapper/app";
+  import Nav from './../components/Nav.svelte'
+  import Hero from './../components/Hero.svelte'
+  import Banners from './../components/Banners.svelte'
+  import Deals from './../components/Deals.svelte'
+  import Carousel from './../components/Carousel.svelte'
+  import ProductSkeleton from './../components/ProductSkeleton.svelte'
+  import { get, put, post } from './../lib/api'
+  import { host } from './../config'
+  import { goto } from '@sapper/app'
   let loading = false,
     latestProducts = [],
     HPProducts = [],
@@ -21,117 +21,117 @@
     R64GBProducts = [],
     R32GBProducts = [],
     R16GBProducts = [],
-    R8GBProducts = [];
+    R8GBProducts = []
 
   async function getLatest() {
     try {
-      const p = await get("electronics");
-      latestProducts = p.data;
-      return latestProducts;
+      const p = await get('electronics')
+      latestProducts = p.data
+      return latestProducts
     } catch (e) {}
   }
   // getLatest();
   async function getHP() {
     try {
-      const p = await get("electronics", { brandName: "HP" });
-      HPProducts = p.data;
-      return HPProducts;
+      const p = await get('electronics', { brandName: 'HP' })
+      HPProducts = p.data
+      return HPProducts
     } catch (e) {}
   }
   // getHP();
   async function getDell() {
     try {
-      const p = await get("electronics", { brandName: "Dell" });
-      DellProducts = p.data;
-      return DellProducts;
+      const p = await get('electronics', { brandName: 'Dell' })
+      DellProducts = p.data
+      return DellProducts
     } catch (e) {}
   }
   // getDell();
   async function getLenovo() {
     try {
-      const p = await get("electronics", { brandName: "Lenovo" });
-      LenovoProducts = p.data;
-      return LenovoProducts;
+      const p = await get('electronics', { brandName: 'Lenovo' })
+      LenovoProducts = p.data
+      return LenovoProducts
     } catch (e) {}
   }
   // getLenovo();
   async function getAsus() {
     try {
-      const p = await get("electronics", { brandName: "Asus" });
-      asusProducts = p.data;
-      return asusProducts;
+      const p = await get('electronics', { brandName: 'Asus' })
+      asusProducts = p.data
+      return asusProducts
     } catch (e) {}
   }
   // getAsus();
   async function getAcer() {
     try {
-      const p = await get("electronics", { brandName: "Acer" });
-      AcerProducts = p.data;
-      return AcerProducts;
+      const p = await get('electronics', { brandName: 'Acer' })
+      AcerProducts = p.data
+      return AcerProducts
     } catch (e) {}
   }
   // getAcer();
   async function getI7() {
     try {
-      const p = await get("electronics/es", { "Processor Name": "Core i7" });
-      I7Products = p.data;
-      return I7Products;
+      const p = await get('electronics/es', { 'Processor Name': 'Core i7' })
+      I7Products = p.data
+      return I7Products
     } catch (e) {}
   }
   // getI7();
   async function getI5() {
     try {
-      const p = await get("electronics/es", { "Processor Name": "Core i5" });
-      I5Products = p.data;
-      return I5Products;
+      const p = await get('electronics/es', { 'Processor Name': 'Core i5' })
+      I5Products = p.data
+      return I5Products
     } catch (e) {}
   }
   // getI5();
   async function getI3() {
     try {
-      const p = await get("electronics/es", { "Processor Name": "Core i3" });
-      I3Products = p.data;
-      return I3Products;
+      const p = await get('electronics/es', { 'Processor Name': 'Core i3' })
+      I3Products = p.data
+      return I3Products
     } catch (e) {}
   }
   // getI3();
   async function getSSD() {
     try {
-      const p = await get("electronics/es", { SSD: "Yes" });
-      SSDProducts = p.data;
-      return SSDProducts;
+      const p = await get('electronics/es', { SSD: 'Yes' })
+      SSDProducts = p.data
+      return SSDProducts
     } catch (e) {}
   }
   // getSSD();
   async function get64GB() {
     try {
-      const p = await get("electronics/es", { RAM: "64 GB" });
-      R64GBProducts = p.data;
-      return R64GBProducts;
+      const p = await get('electronics/es', { RAM: '64 GB' })
+      R64GBProducts = p.data
+      return R64GBProducts
     } catch (e) {}
   }
   // get64GB();
   async function get32GB() {
     try {
-      const p = await get("electronics/es", { RAM: "32 GB" });
-      R32GBProducts = p.data;
-      return R32GBProducts;
+      const p = await get('electronics/es', { RAM: '32 GB' })
+      R32GBProducts = p.data
+      return R32GBProducts
     } catch (e) {}
   }
   // get32GB();
   async function get16GB() {
     try {
-      const p = await get("electronics/es", { RAM: "16 GB" });
-      R16GBProducts = p.data;
-      return R16GBProducts;
+      const p = await get('electronics/es', { RAM: '16 GB' })
+      R16GBProducts = p.data
+      return R16GBProducts
     } catch (e) {}
   }
   // get16GB();
   async function get8GB() {
     try {
-      const p = await get("electronics/es", { RAM: "8 GB" });
-      R8GBProducts = p.data;
-      return R8GBProducts;
+      const p = await get('electronics/es', { RAM: '8 GB' })
+      R8GBProducts = p.data
+      return R8GBProducts
     } catch (e) {}
   }
   // get8GB();
@@ -150,17 +150,17 @@
     name="og_title"
     property="og:title"
     content="Online store for electronics using Litekart ecommerce" />
-  <meta name="og_url" property="og:url" content={`${HOST}/search`} />
+  <meta name="og_url" property="og:url" content={`${host}/search`} />
   <!-- Twitter -->
   <meta name="twitter:title" content="Hand picked laptops for you" />
   <meta
     name="twitter:description"
     content="Online electronics store using Litekart ecommerce platform.
     Features include search, filter, lazy load, infinite scrolling" />
-  <link rel="canonical" href={`${HOST}`} />
-  <link rel="alternate" hreflang="en" href={`${HOST}`} />
+  <link rel="canonical" href={`${host}`} />
+  <link rel="alternate" hreflang="en" href={`${host}`} />
 </svelte:head>
-<Header home={true} />
+<Nav home={true} />
 <!-- <Hero /> -->
 <!-- <Banners /> -->
 {#await getHP()}
