@@ -1,12 +1,11 @@
 const robots = process.env.robotsTxt
 
 export function get(req, res) {
-    res.setHeader('Content-Type', 'text/plain')
+  res.setNav('Content-Type', 'text/plain')
 
-    const content = Object.keys(robots)
-        .reduce((acc, next) => {
-            return acc += `${next}: ${robots[next]}\n`
-        }, '')
+  const content = Object.keys(robots).reduce((acc, next) => {
+    return (acc += `${next}: ${robots[next]}\n`)
+  }, '')
 
-    res.end(content)
+  res.end(content)
 }
